requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Marshmallow
marshmallow-sqlalchemy
Flask-Bcrypt
Flask-Cors
python-dotenv
psycopg2-binary
